+++ tests/NGram.spec.ts
@@ -29,19 +29,19 @@ describe('NGram', () => {
 	});
 
 	describe('exceptions', () => {
-		it('s1 is null', () => {
+		it('s0 is null', () => {
 			try {
 				instance.distance(null as any, 'kitten');
 			} catch (error) {
-				expect(error.message).toBe('s1 must neither be null nor undefined');
+				expect(error.message).toBe('s0 must neither be null nor undefined');
 			}
 		});
 
-		it('s2 is null', () => {
+		it('s1 is null', () => {
 			try {
 				instance.distance('kitten', null as any);
 			} catch (error) {
-				expect(error.message).toBe('s2 must neither be null nor undefined');
+				expect(error.message).toBe('s1 must neither be null nor undefined');
 			}
 		});
 	});